fix: fixed types, linting
