+++ src/core/server/app/handlers/api/index.ts
@@ -1,9 +1,11 @@
 export * from "./account";
 export * from "./auth";
+export * from "./dashboard";
+export * from "./externalMedia";
 export * from "./health";
 export * from "./install";
-export * from "./version";
-export * from "./dashboard";
-export * from "./user";
-export * from "./story";
+export * from "./oembed";
 export * from "./remoteMedia";
+export * from "./story";
+export * from "./user";
+export * from "./version";
