+++ src/core/server/app/router/api/index.ts
@@ -2,8 +2,7 @@ import express from "express";
 import passport from "passport";
 
 import { AppOptions } from "coral-server/app";
-import { externalMediaHandler } from "coral-server/app/handlers/api/externalMedia/externalMedia";
-import { oembedHandler } from "coral-server/app/handlers/api/oembed/oembed";
+import { externalMediaHandler, oembedHandler } from "coral-server/app/handlers";
 import {
   apolloGraphQLMiddleware,
   authenticate,
@@ -56,13 +55,6 @@ export function createAPIRouter(app: AppOptions, options: RouterOptions) {
   router.use("/account", createNewAccountRouter(app, options));
   router.use("/user", createNewUserRouter(app));
 
-  router.get("/oembed", cspSiteMiddleware(app), oembedHandler(app));
-  router.get(
-    "/external-media",
-    cspSiteMiddleware(app),
-    externalMediaHandler(app)
-  );
-
   // Configure the GraphQL route middleware.
   router.use(
     "/graphql",
@@ -87,6 +79,8 @@ export function createAPIRouter(app: AppOptions, options: RouterOptions) {
     loggedInMiddleware,
     createRemoteMediaRouter(app)
   );
+  router.get("/oembed", cspSiteMiddleware(app), oembedHandler(app));
+  router.get("/external-media", cspSiteMiddleware(app), externalMediaHandler());
 
   // General API error handler.
   router.use(notFoundMiddleware);
