+++ src/core/server/graph/resolvers/MediaConfiguration.ts
@@ -9,5 +9,5 @@ export const MediaConfiguration: GQLMediaConfigurationTypeResolver<Partial<
   twitter: ({ twitter = {} }) => twitter,
   youtube: ({ youtube = {} }) => youtube,
   giphy: ({ giphy = {} }) => giphy,
-  external: ({ external = {} }) => external,
+  external: () => ({}),
 };
