+++ src/core/server/models/tenant/helpers.ts
@@ -71,3 +71,19 @@ export function getWebhookEndpoint(
 ) {
   return tenant.webhooks.endpoints.find((e) => e.id === endpointID) || null;
 }
+
+export function supportsMediaType(
+  tenant: Pick<Tenant, "media" | "featureFlags">,
+  type: "twitter" | "youtube" | "giphy" | "external"
+): tenant is Omit<Tenant, "media"> & Required<Pick<Tenant, "media">> {
+  switch (type) {
+    case "external":
+      return hasFeatureFlag(tenant, GQLFEATURE_FLAG.EXTERNAL_MEDIA);
+    case "twitter":
+      return !!tenant.media?.twitter.enabled;
+    case "youtube":
+      return !!tenant.media?.youtube.enabled;
+    case "giphy":
+      return !!tenant.media?.giphy.enabled && !!tenant.media.giphy.key;
+  }
+}
