+++ src/core/server/models/tenant/tenant.ts
@@ -468,19 +468,3 @@ export function retrieveAnnouncementIfEnabled(
   }
   return null;
 }
-
-export function supportsMediaType(
-  tenant: Tenant,
-  type: "twitter" | "youtube" | "giphy" | "external"
-): tenant is Omit<Tenant, "media"> & Required<Pick<Tenant, "media">> {
-  switch (type) {
-    case "external":
-      return !!tenant.media?.external.enabled;
-    case "twitter":
-      return !!tenant.media?.twitter.enabled;
-    case "youtube":
-      return !!tenant.media?.youtube.enabled;
-    case "giphy":
-      return !!tenant.media?.giphy.enabled && !!tenant.media.giphy.key;
-  }
-}
